Weird GIT, hopefully all of Micha's fixes for Linux got in after doing it 'by hand' oh lawd.

+++ src/Common/CVar.cpp
@@ -162,7 +162,7 @@ static void parse_string_value(cvar_t *var)
     if (s[0] == '0' && s[1] == 'x') {
         long v = strtol(s, NULL, 16);
 
-        var->integer = clamp(v, INT_MIN, INT_MAX);
+        var->integer = clamp(v, static_cast<long int>(INT_MIN), static_cast<long int>(INT_MAX));
         var->value = (float)var->integer;
     } else {
         var->integer = atoi(var->string);
